chore: move contact API to serverless function
