+++ src/pages/ContactPage.tsx
@@ -47,7 +47,7 @@ const ContactPage: React.FC = () => {
     setErrors({});
 
     try {
-      const response = await fetch('/api/contact', {
+      const response = await fetch('/.netlify/functions/contact', {
         method: 'POST',
         headers: { 'Content-Type': 'application/json' },
         body: JSON.stringify(formData),
